:beetle: Fix linux build.

+++ src/cs-scene/Trajectory.cpp
@@ -97,8 +97,8 @@ void Trajectory::upload(glm::dmat4 const& relativeTransform, double dTime,
     // transform all points to observer centric coordinates
     std::vector<glm::vec4> points(vPoints.size());
 
-    for (int i(0); i < vPoints.size(); ++i) {
-      int ringbufferIndex = (i + startIndex) % static_cast<int>(vPoints.size());
+    for (size_t i(0); i < vPoints.size(); ++i) {
+      int ringbufferIndex = (static_cast<int>(i) + startIndex) % static_cast<int>(vPoints.size());
 
       glm::dvec4 const& curr = vPoints[ringbufferIndex];
